Flux Iteration modification
Change to the flux iteration step to catch for when an additional
convolution may not be required. Additionally, ensures that iterated
flux measurement does not do an extra convolution of the already
convolved apertures, and thus deblend apertures based on an incorrect
PSF. This previously was possible when using a particular set of input
parameters.