Update Pixel Flux-Weighting for improved behaviour with bad skies
Modify Pixel Flux-Weighting to behave better in cases where the sky
pedestal of the input image is very wrong. Previously, if the sky was
highly variable across the image, and simultaneously highly non-zero,
then pixel flux weighting would cause sky-dependent flux weighting to
occur. This is now combatted, but not entirely removed, by using the
image mode+MAD as the lower bound for the pixel flux weighting.
Furthermore, pixel flux-weighting now can down-weight to a minimum of
1%, which is more conservative than the 0.01% previously allowed.