Removal of useless calculations, incorporation of additional error terms
Removal of a series of calculations that are legacies of long past
testing of the program, which are not used and are not output, thus
having no purpose. Furthermore, only calculate some diagnostic values
when diagnostic is requested, cutting down on unneeded calculations.
Additionally, have added incorporation of Blanks Error and Sky RMS
error to final output uncertainties, so that this value now reflects
what we believe is “Best Practice” uncertainty.