Modified Randoms/Blanks Correction
Modified Randoms/Blanks correction calculation to now perform
additional weighted mean based on object masking. Standard deviation is
now correctly represented using the standard deviation of a weighted
sample, and is implemented for both sd() and mad() functions.
Additionally, fixed a bug in the labels when plotting a sample of
randoms/blanks